* split all src/engine classes into one class per file
