+++ src/engine/BMAttack.php
@@ -3,6 +3,12 @@
 require_once 'BMUtility.php';
 require_once 'BMDie.php';
 
+// require each type of attack
+require_once 'BMAttackPower.php';
+require_once 'BMAttackShadow.php';
+require_once 'BMAttackSkill.php';
+require_once 'BMAttackPass.php';
+
 /**
  * BMAttack: attack validation and commital code.
  *
@@ -186,8 +192,8 @@ class BMAttack {
         }
 
         // OK, these aren't necessary for this one, but it's consistent.
-        $aIt = new XCYIterator($attackers, 1);
-        $dIt = new XCYIterator($defenders, 1);
+        $aIt = new BMUtilityXCYIterator($attackers, 1);
+        $dIt = new BMUtilityXCYIterator($defenders, 1);
 
         foreach ($aIt as $att) {
             foreach ($dIt as $def) {
@@ -211,7 +217,7 @@ class BMAttack {
 
         $count = count($many);
 
-        $oneIt = new XCYIterator($one, 1);
+        $oneIt = new BMUtilityXCYIterator($one, 1);
 
         $checkedSizes = array();
 
@@ -229,7 +235,7 @@ class BMAttack {
             // can search the complement of the set at the same time.
             $checkedSizes[$count - $i] = TRUE;
 
-            $manyIt = new XCYIterator($many, $i);
+            $manyIt = new BMUtilityXCYIterator($many, $i);
 
             foreach ($manyIt as $m) {
                 foreach ($oneIt as $o) {
@@ -293,233 +299,5 @@ class BMAttack {
         return $helpers;
     }
 }
-
-
-class BMAttackPower extends BMAttack {
-    public $name = "Power";
-    public $type = 'Power';
-
-    public function find_attack($game) {
-        $targets = $game->defenderAllDieArray;
-
-        return $this->search_onevone($game, $this->validDice, $targets);
-    }
-
-    public function validate_attack($game, $attackers, $defenders) {
-        if (count($attackers) != 1 || count($defenders) != 1) {
-            return FALSE;
-        }
-
-        $helpers = $this->collect_helpers($game, $attackers, $defenders);
-
-        $bounds = $this->help_bounds($helpers);
-
-        foreach ($attackers[0]->attack_values($this->type) as $aVal) {
-
-            if ($aVal + $bounds[1] >= $defenders[0]->defense_value($this->type)) {
-
-                if ($attackers[0]->valid_attack($this->type, $attackers, $defenders) &&
-                    $defenders[0]->valid_target($this->type, $attackers, $defenders))
-                {
-                    return TRUE;
-                }
-            }
-
-        }
-
-        return FALSE;
-    }
-
-    // return how much help is needed and who can contribute
-    public function calculate_contributions($game, $attackers, $defenders) {
-        return array(0, array());
-    }
-}
-
-class BMAttackSkill extends BMAttack {
-    public $name = "Skill";
-    public $type = "Skill";
-
-    // Especially once you take stinger and constant into account,
-    // searching the full attack space is slow and complex
-    //
-    // Building a hit table once trades some increased setup cost for
-    // a much reduced search cost
-    //
-    // Fire still makes life more complex than it might be.
-    //
-    // "Premature optimization is the root of all evil." -- Knuth
-    protected $hit_table = NULL;
-
-    public function find_attack($game) {
-        if (!$this->hit_table) {
-            $this->hit_table = new BMHitTable($this->validDice);
-        }
-
-        $targets = $game->defenderAllDieArray;
-
-        if (count($targets) < 1) { return FALSE; }
-
-        // Check all precise hits before trying any with help, because
-        // help is slow
-        foreach ($targets as $def) {
-            if ($this->hit_table->find_hit($def->defense_value($this->type))) {
-                return TRUE;
-            }
-        }
-
-        // Potentially save some time by checking for the possibility of help
-
-        $help_found = FALSE;
-        foreach ($targets as $def) {
-            foreach ($game->attackerAllDieArray as $att) {
-                if ($this->collect_helpers($game, array($att), array($def))) {
-                    $help_found = TRUE;
-                    break 2;
-                }
-            }
-        }
-
-        if (!$help_found) { return FALSE; }
-
-        $hits = $this->hit_table->list_hits();
-        sort($hits);
-
-        // Should perhaps start around the defending die's value and
-        // work outward, but probably not worth the extra overhead to
-        // do so. We half-ass it by starting in the middle.
-
-        // PHP, why do you have no integer division operator?
-        $i = (int)(count($hits) / 2);
-        $j = $i - 1;
-        $j += count($hits) % 2; // Start in the exact middle of an odd count
-        while ($j >= 0) {
-            foreach ($targets as $t) {
-                $def = array($t);
-                $combos = $this->hit_table->find_hit($hits[$i]);
-                foreach ($combos as $att) {
-                    if ($this->validate_attack($game, $att, $def)) {
-                        return TRUE;
-                    }
-                }
-                if ($i == $j) { continue; }
-                $combos = $this->hit_table->find_hit($hits[$j]);
-                foreach ($combos as $att) {
-                    if ($this->validate_attack($game, $att, $def)) {
-                        return TRUE;
-                    }
-                }
-            }
-            $j--; $i++;
-        }
-
-        return FALSE;
-    }
-
-    public function validate_attack($game, $attackers, $defenders) {
-        if (count($attackers) < 1 || count($defenders) != 1) {
-            return FALSE;
-        }
-
-        // array_intersect tries to convert to strings, so we
-        // use array_uintersect, which needs a comparison
-        // function
-        $cmp = function ($v1,$v2) {
-            if ($v1===$v2) { return 0; }
-            if ($v1 > $v2) { return 1; }
-            return -1;
-        };
-
-        $dval = $defenders[0]->defense_value($this->type);
-
-        // exact hits
-        $combos = $this->hit_table->find_hit($dval);
-        if ($combos) {
-            foreach ($combos as $c) {
-                if (count($c) == count($attackers) &&
-                    count(array_uintersect($c, $attackers, $cmp)) ==
-                    count($c)) {
-                    return TRUE;
-                }
-            }
-        }
-
-        // assisted attacks
-        $helpers = $this->collect_helpers($game, $attackers, $defenders);
-        $bounds = $this->help_bounds($helpers);
-        if ($bounds[0] == 0 && $bounds[1] == 0) { return FALSE; }
-        for ($i = $bounds[0]; $i <= $bounds[1]; $i++) {
-            $combos = $this->hit_table->find_hit($dval + $i);
-            if ($combos) {
-                foreach ($combos as $c) {
-                    if (count($c) == count($attackers) &&
-                        count(array_uintersect($c, $attackers, $cmp)) ==
-                        count($c)) {
-                        return TRUE;
-                    }
-                }
-            }
-        }
-        return FALSE;
-    }
-
-    // return how much help is needed and who can contribute
-    public function calculate_contributions($game, $attackers, $defenders) {
-        return array(0, array());
-    }
-}
-
-
-class BMAttackShadow extends BMAttackPower {
-    public $name = "Shadow";
-    public $type = "Shadow";
-
-    public function validate_attack($game, $attackers, $defenders) {
-        if (count($attackers) != 1 || count($defenders) != 1) {
-            return FALSE;
-        }
-
-
-        $helpers = $this->collect_helpers($game, $attackers, $defenders);
-
-        $bounds = $this->help_bounds($helpers);
-
-        foreach ($attackers[0]->attack_values($this->type) as $aVal) {
-
-            if ($defenders[0]->defense_value($this->type) <= $attackers[0]->max &&
-                ($defenders[0]->defense_value($this->type) >= $aVal ||
-                 $defenders[0]->defense_value($this->type) >= $aVal + $bounds[0])) {
-
-                if ($attackers[0]->valid_attack($this->type, $attackers, $defenders) &&
-                    $defenders[0]->valid_target($this->type, $attackers, $defenders))
-                {
-                    return TRUE;
-                }
-            }
-        }
-
-        return FALSE;
-    }
-
-    // return how much help is needed and who can contribute
-    public function calculate_contributions($game, $attackers, $defenders) {
-        return array(0, array());
-    }
-}
-
-class BMAttackPass extends BMAttack {
-    public $name = "Pass";
-    public $type = "Pass";
-
-    public function find_attack($game) {
-        return $this->validate_attack($game, $this->validDice,
-                                      $game->defenderAttackDieArray);
-    }
-
-    public function validate_attack($game, $attackers, $defenders) {
-        return (empty($attackers) && empty($defenders));
-    }
-}
-
 
 ?>
