+++ src/engine/BMAttackPass.php
@@ -1,0 +1,21 @@
+<?php
+
+require_once 'BMUtility.php';
+require_once 'BMDie.php';
+
+class BMAttackPass extends BMAttack {
+    public $name = "Pass";
+    public $type = "Pass";
+
+    public function find_attack($game) {
+        return $this->validate_attack($game, $this->validDice,
+                                      $game->defenderAttackDieArray);
+    }
+
+    public function validate_attack($game, $attackers, $defenders) {
+        return (empty($attackers) && empty($defenders));
+    }
+}
+
+
+?>
