+++ src/engine/BMContainerInitial.php
@@ -1,0 +1,11 @@
+<?php
+
+// This class may not be needed; the engine could possibly use
+// generic containers and simply open them first.
+
+// Special container for Auxiliary dice and the like. Activated only
+// at the start of games.
+class BMInitialContainer extends BMContainer {
+}
+
+?>
