+++ src/engine/BMContainerPlasma.php
@@ -1,0 +1,6 @@
+<?php
+
+class BMContainerPlasma extends BMContainer {
+}
+
+?>
