+++ src/engine/BMContainerSelect.php
@@ -1,0 +1,6 @@
+<?php
+
+class BMContainerSelect extends BMContainer {
+}
+
+?>
