+++ src/engine/BMContainerSideboard.php
@@ -1,0 +1,11 @@
+<?php
+
+// This class may not be needed; the engine could possibly use
+// generic containers and simply open them last.
+
+// Special container that is opened last each round, so the dice
+// within can replace other dice.
+class BMSideboardContainer extends BMContainer {
+}
+
+?>
