+++ src/engine/BMDieOption.php
@@ -1,0 +1,6 @@
+<?php
+
+class BMDieOption extends BMDie {
+}
+
+?>
