+++ src/engine/BMDieTwin.php
@@ -1,0 +1,6 @@
+<?php
+
+class BMDieTwin extends BMDie {
+}
+
+?>
