+++ src/engine/BMDieWildcard.php
@@ -1,0 +1,6 @@
+<?php
+
+class BMDieWildcard extends BMDie {
+}
+
+?>
