+++ test/src/engine/AttackTests/BMAttackShadowTest.php
@@ -1,6 +1,7 @@
 <?php
 
 require_once "engine/BMAttack.php";
+require_once "engine/BMAttackShadow.php";
 
 /**
  * Generated by PHPUnit_SkeletonGenerator 1.2.0 on 2012-12-21 at 15:12:50.
