+++ test/src/engine/BMDieSwingTest.php
@@ -5,9 +5,9 @@ require_once "engine/BMGame.php";
 require_once "testdummies.php";
 
 
-class BMSwingDieTest extends PHPUnit_Framework_TestCase {
+class BMDieSwingTest extends PHPUnit_Framework_TestCase {
     /**
-     * @var BMSwingDie
+     * @var BMDieSwing
      */
     protected $object;
 
@@ -16,7 +16,7 @@ class BMSwingDieTest extends PHPUnit_Framework_TestCase {
      * This method is called before a test is executed.
      */
     protected function setUp() {
-        $this->object = new BMSwingDie;
+        $this->object = new BMDieSwing;
     }
 
     /**
@@ -28,7 +28,7 @@ class BMSwingDieTest extends PHPUnit_Framework_TestCase {
     }
 
     /**
-     * @covers BMSwingDie::swing_range
+     * @covers BMDieSwing::swing_range
      */
     public function testSwing_range() {
         foreach (str_split("RSTUVWXYZ") as $swing) {
@@ -48,7 +48,7 @@ class BMSwingDieTest extends PHPUnit_Framework_TestCase {
 
     /**
      * @depends testSwing_range
-     * @covers BMSwingDie::init
+     * @covers BMDieSwing::init
      */
     public function testInit () {
         $this->object->init("X", array());
@@ -114,16 +114,16 @@ class BMSwingDieTest extends PHPUnit_Framework_TestCase {
 
     /**
      * @depends testInit
-     * @covers BMSwingDie::create
+     * @covers BMDieSwing::create
      */
     public function testCreate() {
 
         // Test all valid swing types
         foreach (str_split("RSTUVWXYZ") as $swing) {
-            $die = BMSwingDie::create($swing, array());
+            $die = BMDieSwing::create($swing, array());
 
             $this->assertInstanceOf('BMDie', $die);
-            $this->assertInstanceOf('BMSwingDie', $die);
+            $this->assertInstanceOf('BMDieSwing', $die);
             $this->assertEquals($swing, $die->swingType);
         }
 
@@ -132,7 +132,7 @@ class BMSwingDieTest extends PHPUnit_Framework_TestCase {
         foreach (str_split("rstuvwxyzaQ") as $swing) {
             $fail = FALSE;
             try {
-                $die = BMSwingDie::create($swing, array());
+                $die = BMDieSwing::create($swing, array());
             } catch (UnexpectedValueException $e) {
                 $fail = TRUE;
             }
@@ -144,7 +144,7 @@ class BMSwingDieTest extends PHPUnit_Framework_TestCase {
 
         // try some more bad values
         try {
-            $die = BMSwingDie::create(6, array());
+            $die = BMDieSwing::create(6, array());
         }
         catch (UnexpectedValueException $e) {
             $fail = TRUE;
@@ -154,7 +154,7 @@ class BMSwingDieTest extends PHPUnit_Framework_TestCase {
         $fail = FALSE;
 
         try {
-            $die = BMSwingDie::create("RT", array());
+            $die = BMDieSwing::create("RT", array());
         }
         catch (UnexpectedValueException $e) {
             $fail = TRUE;
@@ -164,7 +164,7 @@ class BMSwingDieTest extends PHPUnit_Framework_TestCase {
         $fail = FALSE;
 
         try {
-            $die = BMSwingDie::create("0.A", array());
+            $die = BMDieSwing::create("0.A", array());
         }
         catch (UnexpectedValueException $e) {
             $fail = TRUE;
@@ -177,7 +177,7 @@ class BMSwingDieTest extends PHPUnit_Framework_TestCase {
 
     /**
      * @depends testInit
-     * @covers BMSwingDie::activate
+     * @covers BMDieSwing::activate
      */
 
     public function testActivate () {
@@ -218,7 +218,7 @@ class BMSwingDieTest extends PHPUnit_Framework_TestCase {
 
     /**
      * @depends testInit
-     * @covers BMSwingDie::split
+     * @covers BMDieSwing::split
      */
     public function testSplit() {
         $this->object->init("X");
@@ -326,7 +326,7 @@ class BMSwingDieTest extends PHPUnit_Framework_TestCase {
     /**
      * @depends testInit
      * @depends testSplit
-     * @covers BMSwingDie::set_swingValue
+     * @covers BMDieSwing::set_swingValue
      */
     public function testSet_swingValue() {
 
@@ -435,7 +435,7 @@ class BMSwingDieTest extends PHPUnit_Framework_TestCase {
      * @depends testInit
      * @depends testActivate
      * @depends testSet_swingValue
-     * @covers BMSwingDie::roll
+     * @covers BMDieSwing::roll
      */
     public function testRoll() {
 
@@ -601,7 +601,7 @@ class BMSwingDieTest extends PHPUnit_Framework_TestCase {
      * @depends testInit
      * @depends testSet_swingValue
      * @depends testRoll
-     * @covers BMSwingDie::make_play_die
+     * @covers BMDieSwing::make_play_die
      */
     public function testMake_play_die() {
         $game = new DummyGame;
@@ -697,7 +697,7 @@ class BMSwingDieTest extends PHPUnit_Framework_TestCase {
     }
 
     /*
-     * @covers BMSwingDie::describe
+     * @covers BMDieSwing::describe
      */
     public function testDescribe() {
         // Remove the following lines when you implement this test.
