+++ test/src/engine/Utility/HitTableTest.php
@@ -2,10 +2,10 @@
 /**
  * Generated by PHPUnit_SkeletonGenerator 1.2.0 on 2013-01-29 at 13:10:57.
  */
-class BMHitTableTest extends PHPUnit_Framework_TestCase
+class HitTableTest extends PHPUnit_Framework_TestCase
 {
     /**
-     * @var BMHitTable
+     * @var BMUtilityHitTable
      */
     protected $object;
 
@@ -24,7 +24,7 @@ class BMHitTableTest extends PHPUnit_Framework_TestCase
         $die4 = BMDie::create(20);
         $die4->value = 18;
 
-        $this->object = new BMHitTable(array($die1, $die2, $die3, $die4));
+        $this->object = new BMUtilityHitTable(array($die1, $die2, $die3, $die4));
     }
 
     /**
@@ -36,8 +36,8 @@ class BMHitTableTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers BMHitTable::find_hit
-     * @covers BMHitTable::__construct
+     * @covers BMUtilityHitTable::find_hit
+     * @covers BMUtilityHitTable::__construct
      */
     public function testFind_hit()
     {
@@ -104,7 +104,7 @@ class BMHitTableTest extends PHPUnit_Framework_TestCase
         $die3 = BMDie::create(3);
         $die3->value = 1;
 
-        $ht = new BMHitTable(array($die1, $die2, $die3));
+        $ht = new BMUtilityHitTable(array($die1, $die2, $die3));
 
         $ret = $ht->find_hit(1);
         $this->assertEquals(3, count($ret));
@@ -154,7 +154,7 @@ class BMHitTableTest extends PHPUnit_Framework_TestCase
         $die2 = BMDie::create(6);
         $die2->value = 6;
 
-        $ht = new BMHitTable(array($die1, $die2));
+        $ht = new BMUtilityHitTable(array($die1, $die2));
 
         // Two ways to make 6-10, one for all the rest
         foreach (array(6, 7, 8, 9, 10) as $i) {
@@ -207,7 +207,7 @@ class BMHitTableTest extends PHPUnit_Framework_TestCase
         $die2 = BMDie::create(6, array("TestStinger"));
         $die2->value = 6;
 
-        $ht = new BMHitTable(array($die1, $die2));
+        $ht = new BMUtilityHitTable(array($die1, $die2));
 
         // One way to make 11-16
         foreach (array(11, 12, 13, 14, 15, 16) as $i) {
@@ -264,8 +264,8 @@ class BMHitTableTest extends PHPUnit_Framework_TestCase
     }
 
     /**
-     * @covers BMHitTable::list_hits
-     * @covers BMHitTable::__construct
+     * @covers BMUtilityHitTable::list_hits
+     * @covers BMUtilityHitTable::__construct
      */
     public function testList_hits()
     {
@@ -290,7 +290,7 @@ class BMHitTableTest extends PHPUnit_Framework_TestCase
         $die3 = BMDie::create(3);
         $die3->value = 1;
 
-        $ht = new BMHitTable(array($die1, $die2, $die3));
+        $ht = new BMUtilityHitTable(array($die1, $die2, $die3));
 
         for ($i = 1; $i <= 3; $i++) {
             $this->assertContains($i, $ht->list_hits());
@@ -305,7 +305,7 @@ class BMHitTableTest extends PHPUnit_Framework_TestCase
         $die2->value = 6;
 
 
-        $ht = new BMHitTable(array($die1, $die2));
+        $ht = new BMUtilityHitTable(array($die1, $die2));
 
         for ($i = -64; $i < 64; $i++) {
             // They should cover 1-4, 6-10
@@ -322,7 +322,7 @@ class BMHitTableTest extends PHPUnit_Framework_TestCase
         $die2 = BMDie::create(8, array("TestStinger"));
         $die2->value = 8;
 
-        $ht = new BMHitTable(array($die1, $die2));
+        $ht = new BMUtilityHitTable(array($die1, $die2));
 
         for ($i = -64; $i < 64; $i++) {
             // They should cover 1-12
