+++ test/src/engine/testdummies.php
@@ -4,6 +4,7 @@ require_once "engine/BMDie.php";
 require_once "engine/BMSkill.php";
 require_once "engine/BMContainer.php";
 require_once "engine/BMAttack.php";
+require_once "engine/BMAttackSkill.php";
 
 // well defined skill classes with which to test
 class BMSkillTesting extends BMSkill {
@@ -123,7 +124,7 @@ class BMAttSkillTesting extends BMAttackSkill {
     }
 
     public function make_hit_table() {
-        $this->hit_table = new BMHitTable($this->validDice);
+        $this->hit_table = new BMUtilityHitTable($this->validDice);
     }
 }
